requirements.txt
google-genai
pandas
python-dotenv
pydantic
rich